% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfit.R
\name{feature_enrichment}
\alias{feature_enrichment}
\title{Test features for enrichment}
\usage{
feature_enrichment(p)
}
\arguments{
\item{p}{A \code{pfit} object}
}
\value{
A data frame with three columns, giving the observation, the
  feature, and the p-value for enrichment.
}
\description{
Tests each feature for enrichment. The null hypothesis is that all the
features across all observations arise from a single Dirichlet-Multinomial
distribution.
}
