% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lungreplicate_reads}
\alias{lungreplicate_reads}
\title{Lung microbiome replicates data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19362 rows and 3 columns.
}
\usage{
lungreplicate_reads
}
\description{
Lung microbiome replicates data
}
\keyword{datasets}
