\name{pbetabinom}
\alias{pbetabinom}
\title{Beta-binomial distribution: cumulative distribution function}
\usage{
  pbetabinom(k, n, a, b, log.p = TRUE)
}
\arguments{
  \item{k}{Vector of successes.}

  \item{n}{Number of trials.}

  \item{a}{First shape parameter of Beta distribution (must
  be positive)}

  \item{b}{Second shape parameter of Beta distribution
  (must be positive)}

  \item{log.p}{If TRUE, return the natural logarithm}
}
\value{
  The cumulative probability of observing k successes in n
  trials
}
\description{
  Beta-binomial distribution: cumulative distribution
  function
}

