% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfit.R
\name{pfit_enrichment}
\alias{pfit_enrichment}
\title{Test features for enrichment}
\usage{
pfit_enrichment(p)
}
\arguments{
\item{p}{A \code{pfit} object}
}
\value{
A data frame with three columns:
  \item{observation_idx}{The index of the observation, i.e. the row in the
    input data.}
  \item{feature_idx}{The index of the feature, i.e. the column in the input
    data.}
  \item{p_value}{The p-value for feature enrichment against a null
    hypothesis of all features arising from a single Dirichlet-Multinomial
    distribution.}
}
\description{
Test features for enrichment
}
