% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfit.R
\name{plot.pfit}
\alias{plot.pfit}
\title{Plot a \code{pfit} object}
\usage{
\method{plot}{pfit}(x, ...)
}
\arguments{
\item{x}{A \code{pfit} object}

\item{...}{Arguments passed to other functions}
}
\value{
A ggplot object
}
\description{
Plot a \code{pfit} object
}
