% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polyafit_example_data}
\alias{polyafit_example_data}
\title{Example data for pfit}
\format{
A matrix with three observations and 50 features. Each observation is
generated from a multinomial distribution with the same underlying feature
probabilities.
}
\usage{
polyafit_example_data
}
\description{
Example data for pfit
}
\keyword{datasets}
