\name{optim_polya}
\alias{optim_polya}
\title{Find MLE of a multivariate Polya distribution}
\usage{
  optim_polya(counts)
}
\arguments{
  \item{counts}{A matrix of observed data, one column per
  category, one row per trial.}
}
\value{
  The result of the final call to optim().
}
\description{
  Find MLE of a multivariate Polya distribution
}

