% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfit.R
\name{pfit}
\alias{pfit}
\alias{pfit.matrix}
\alias{pfit.data.frame}
\title{S3 interface to polyafit}
\usage{
pfit(x, min_counts = 5, ...)

\method{pfit}{matrix}(x, min_counts = 5)

\method{pfit}{data.frame}(x, min_counts = 5, rownames_in = 1)
}
\arguments{
\item{x}{A matrix or data frame with one row per observation and column per
feature}

\item{min_counts}{Minimum number of observed counts for a feature to be
included in the fit}

\item{...}{Additional arguments passed to other methods}

\item{rownames_in}{The column of the data frame that gives the rownames
for the data. Can be specified as an integer index or with the column
name in quotes. Set to \code{NULL} to use the existing rownames of the
data frame.}
}
\value{
An object of class "pfit", with the following attributes:
  \item{data}{The data matrix. Row and column names will be created
    automatically if the original data matrix lacked them.}
  \item{is_included}{Boolean vector indicating if the feature was included
    in the model.}
  \item{params}{Best-fit model parameters.}
  \item{theta}{The overdispersion parameter, \code{1 / sum(params)}.}
}
\description{
S3 interface to polyafit
}
\examples{
x <- matrix(c(56, 42, 122, 100, 8, 15, 21, 14, 82, 98), nrow = 2)
p <- pfit(x)
p$params
p$theta
}
